backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.4
langchain==0.2.16
langchain-openai==0.1.23
langgraph==0.2.19
openai==1.40.0
sounddevice==0.4.7
soundfile==0.12.1
